moved computational utils to module
